refactor : enum 타입 이동으로 인한 패키지 주소 변경
