+++ src/main/java/sisosolsol/greenfire/category/model/dto/CategoryUpdateDTO.java
@@ -2,7 +2,7 @@ package sisosolsol.greenfire.category.model.dto;
 
 import lombok.Getter;
 import lombok.ToString;
-import sisosolsol.greenfire.category.model.type.CategoryType;
+import sisosolsol.greenfire.common.enums.CategoryType;
 
 @Getter
 @ToString
